refactor directory hierarchy for two chisel versions
