+++ src/chisel/main/scala/xiangshan/transforms/PrintModuleName.scala
@@ -14,23 +14,27 @@
 * See the Mulan PSL v2 for more details.
 ***************************************************************************************/
 
-package xstransforms
+package xiangshan.transforms
 
 import firrtl._
 import firrtl.ir._
-import firrtl.stage.TransformManager.TransformDependency
 import utils.XSLog
+import firrtl.options.Phase
+import firrtl.stage.FirrtlCircuitAnnotation
 
-class PrintModuleName extends Transform with DependencyAPIMigration {
+class PrintModuleName extends Phase {
 
-  // avoid print's check
-  override def prerequisites = firrtl.stage.Forms.Checks
-  override def invalidates(a: Transform) = false
-  override def optionalPrerequisiteOf: Seq[TransformDependency] = firrtl.stage.Forms.HighEmitters
+  override def invalidates(a: Phase) = false
 
-  override protected def execute(state: CircuitState): CircuitState = {
+  override def transform(annotations: AnnotationSeq): AnnotationSeq = {
 
-    val c = state.circuit
+    import xiangshan.transforms.Helpers._
+
+    val (Seq(circuitAnno: FirrtlCircuitAnnotation), otherAnnos) = annotations.partition {
+      case _: FirrtlCircuitAnnotation => true
+      case _ => false
+    }
+    val c = circuitAnno.circuit
 
     def onStmt(s: Statement): Statement = s match {
       case Print(info, StringLit(string), args, clk, en) =>
@@ -39,8 +43,6 @@ class PrintModuleName extends Transform with DependencyAPIMigration {
         other.mapStmt(onStmt)
     }
 
-    state.copy(c.mapModule(m => m.mapStmt(onStmt)))
-
+    FirrtlCircuitAnnotation(c.mapModule(m => m.mapStmt(onStmt))) +: otherAnnos
   }
-
 }
