+++ src/chisel3/main/scala/xiangshan/transforms/PrintControl.scala
@@ -14,7 +14,7 @@
 * See the Mulan PSL v2 for more details.
 ***************************************************************************************/
 
-package xstransforms
+package xiangshan.transforms
 
 import firrtl._
 import firrtl.ir._
