+++ src/test/scala/fu/IntDiv.scala
@@ -24,7 +24,7 @@ import chiseltest.simulator.VerilatorFlags
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.must.Matchers
 import xiangshan.test.types._
-import xstransforms.PrintModuleName
+import xiangshan.transforms.PrintModuleName
 
 import xiangshan.backend.fu._
